fix: 종료된 스터디에서는 보상 사용 버튼 미노출 구현 #445
